+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   `epy`:
     *   Add a `epy.classproperty`
+    *   Better error when `epy.pretty_repr_top_level` is miss-used.
 *   `ecolab`:
     *   `adhoc`:
         *   Use the official DM colab backend.